dvc-data: isolate hashfile-related logic
With the current discussion around cloud versioning and import handling,
it is clear that dvc-data is not purely about object workflow, but more
about data index operations with hashfile being one of possible backends
for backing up your data (similar to cloud versioning). Thus we need to
cleanly separate these to some extent to reflect our intentions and
high-level logic.
