+++ tests/benchmarks/test_checkout.py
@@ -5,8 +5,8 @@ import pytest
 from dvc_objects.fs import localfs
 from dvc_objects.fs.generic import test_links as _test_links
 
-from dvc_data.checkout import checkout
 from dvc_data.cli import build, gentree, get_odb
+from dvc_data.hashfile.checkout import checkout
 from dvc_data.hashfile.state import State
